refactor: profit을 participants 내부로 통합
