+++ src/test/java/blackjack/domain/participant/ParticipantsTest.java
@@ -1,10 +1,16 @@
 package blackjack.domain.participant;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.assertj.core.api.Assertions.entry;
 
 import java.util.Collections;
 import java.util.List;
+import java.util.Map;
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardRank;
+import blackjack.domain.card.CardShape;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -58,5 +64,31 @@ class ParticipantsTest {
 
         assertThatThrownBy(() -> new Participants(playerNames, invalidSizePlayersMoney))
                 .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @DisplayName("전체 참여자들의 수익금을 계산할 수 있다.")
+    @Test
+    void createAllProfit() {
+        Participants allPlayerBustParticipants = createAllPlayerBustParticipants();
+
+        Map<String, Double> result = allPlayerBustParticipants.createAllProfit();
+
+        assertThat(result)
+                .hasSize(2)
+                .containsExactly(entry("딜러", 1000.0), entry("pobi", -1000.0));
+    }
+
+    private Participants createAllPlayerBustParticipants() {
+        Participants participants = new Participants(
+                List.of("pobi"),
+                List.of(new Money(1000)));
+        List<Player> players = participants.getPlayers();
+        for (Player player : players) {
+            player.hit(new Card(CardRank.QUEEN, CardShape.DIAMOND));
+            player.hit(new Card(CardRank.JACK, CardShape.DIAMOND));
+            player.hit(new Card(CardRank.KING, CardShape.DIAMOND));
+        }
+
+        return participants;
     }
 }
